+++ ipylab/jupyterfrontend.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import asyncio
 from typing import NotRequired, Self, TypedDict
 
-from traitlets import Instance, Unicode
+from traitlets import Dict, Instance, Tuple, Unicode
 
 from ipylab._plugin_manger import pm
 from ipylab.asyncwidget import AsyncWidgetBase, register, widget_serialization
@@ -29,10 +29,13 @@ class JupyterFrontEnd(AsyncWidgetBase):
 
     version = Unicode(read_only=True).tag(sync=True)
     commands = Instance(CommandRegistry, (), read_only=True).tag(sync=True, **widget_serialization)
-    sessions = Instance(SessionManager, (), read_only=True).tag(sync=True, **widget_serialization)
     command_pallet = Instance(CommandPalette, (), read_only=True).tag(
         sync=True, **widget_serialization
     )
+
+    current_widget_id = Unicode(read_only=True).tag(sync=True)
+    current_session = Dict(read_only=True).tag(sync=True)
+    all_sessions = Tuple(read_only=True).tag(sync=True)
 
     @property
     def dialog(self) -> Dialog:
@@ -42,15 +45,21 @@ class JupyterFrontEnd(AsyncWidgetBase):
 
     @property
     def file_dialog(self) -> FileDialog:
-        if not hasattr(self, "_file_dialog"):
-            self._file_dialog = FileDialog()
-        return self._file_dialog
+        if not hasattr(self, "_fileDialog"):
+            self._fileDialog = FileDialog()
+        return self._fileDialog
 
     @property
     def shell(self) -> Shell:
         if not hasattr(self, "_shell"):
             self._shell = Shell()
         return self._shell
+
+    @property
+    def sessionManager(self) -> SessionManager:
+        if not hasattr(self, "_sessionManger"):
+            self._sessionManger = SessionManager()
+        return self._sessionManger
 
     async def wait_ready(self, timeout=5) -> Self:
         """Wait until connected to app indicates it is ready."""
@@ -59,7 +68,6 @@ class JupyterFrontEnd(AsyncWidgetBase):
                 group.create_task(super().wait_ready())
                 group.create_task(self.commands.wait_ready())
                 group.create_task(self.command_pallet.wait_ready())
-                group.create_task(self.sessions.wait_ready())
         return self
 
     def _init_python_backend(self) -> str: